created UserController for handling user requests.
