+++ backend/pronunciationAppBack/src/main/java/dev/pronunciationAppBack/controller/UserController.java
@@ -1,0 +1,44 @@
+package dev.pronunciationAppBack.controller;
+
+import dev.pronunciationAppBack.model.User;
+import dev.pronunciationAppBack.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Date;
+import java.util.Optional;
+
+@RestController
+@RequestMapping("/api/user")
+
+public class UserController {
+    @Autowired
+    private UserRepository userRepository;
+
+    @GetMapping("/user/{id}")
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+        Optional<User> user = userRepository.findById(id);
+        HttpHeaders headers = getCommonHeaders("Get user by ID");
+
+        return user.map(value -> new ResponseEntity<>(value, headers, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(headers, HttpStatus.NOT_FOUND));
+    }
+
+    private HttpHeaders getCommonHeaders(String description) {
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("desc", description);
+        headers.add("content-type", "application/json");
+        headers.add("date", new Date().toString());
+        headers.add("server", "Spring Boot");
+        headers.add("version", "1.0.0");
+        headers.add("word-count", String.valueOf(userRepository.count()));
+        headers.add("object", "words");
+        return headers;
+    }
+}
